% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_active.R
\name{find_active}
\alias{find_active}
\title{Find Active Function}
\usage{
find_active(ind, min)
}
\arguments{
\item{ind}{A list of indices from MJO filtered dataset}

\item{min}{A numeric that is of a certain value and all values above will return}
}
\description{
Takes a list of MJO indices, ind, and a minimum criteria, min,
and returns a vector with the dates of all matches
}
\examples{
find_active(data, 1)
}
