% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_case.R
\name{find_case}
\alias{find_case}
\title{Find Case Study Function}
\usage{
find_case(case, ocur)
}
\arguments{
\item{case}{Numeric list of amount of days separate active events are occuring}

\item{ocurmin}{Amount of occurences to define a signficant MJO}
}
\description{
Uses previous functions to determine from list of active events, case, which are the long enough in days, ocurmin,
and establish which ones are significant enough to further analyze as a case study.
}
\examples{
find_active(data, 30)
}
