% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_ocur.R
\name{find_ocur}
\alias{find_ocur}
\title{Find Occurence Function}
\usage{
find_ocur(ind)
}
\arguments{
\item{ind}{A list of indices from MJO filtered dataset}
}
\description{
Gives active MJO events from a list of indices find the length of days it occurs for each event.
}
\examples{
find_active(data, 30)
}
